+++ _Typings/gameApi/gameApi.d.ts
@@ -21,15 +21,8 @@ declare module "gameApi" {
 
     export { math } from "math"
 
-    export {
-        Float2,
-        Float3,
-        Float4,
-        Quaternion,
-        Bool3,
-        Trans,
-        ColorRGBA,
-    } from "basicData"
+    export { Float2, Float3, Float4, Quaternion, Bool3, Trans } from "basicData"
+    export { ColorRGBA } from "basicData$Game"
 
     /** @deprecated */
     export const __esModule: never